refactor: [#255] move database module into config/core
- Convert core.rs into folder module (config/core/)
- Move database module into config/core/database/
- Update imports to reflect new structure
- Database types now exported through config module hierarchy
- Improves logical grouping (database is part of core config)
- All 1582 tests passing
- All linters passing
